+++ tests/org.hibernate.eclipse.console.test/src/org/hibernate/eclipse/console/test/ConsolePluginAllTests.java
@@ -5,7 +5,6 @@ import java.io.IOException;
 import org.hibernate.eclipse.console.test.mappingproject.MappingTestsAnnotations;
 import org.hibernate.eclipse.console.test.mappingproject.MappingTestsCore;
 import org.hibernate.eclipse.console.test.mappingproject.MappingTestsJpa;
-import org.hibernate.eclipse.mapper.HBMInfoExtractorTest;
 
 import junit.framework.Test;
 import junit.framework.TestSuite;
@@ -19,8 +18,6 @@ public class ConsolePluginAllTests {
 		suite.addTestSuite(MappingTestsCore.class);
 		suite.addTestSuite(MappingTestsJpa.class);
 		suite.addTestSuite(MappingTestsAnnotations.class);
-		
-		suite.addTestSuite(HBMInfoExtractorTest.class);
 		
 		return suite;
 	}
