+++ tests/org.jboss.tools.hibernate.orm.test/META-INF/MANIFEST.MF
@@ -30,7 +30,8 @@ Require-Bundle: org.hibernate.eclipse,
  org.eclipse.jdt.apt.core,
  org.eclipse.datatools.connectivity,
  org.jboss.tools.hibernate.ui,
- org.eclipse.ui.workbench
+ org.eclipse.ui.workbench,
+ org.hibernate.eclipse.mapper
 Bundle-ActivationPolicy: lazy
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Import-Package: org.eclipse.osgi.util
